backend/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
pytrends==4.9.2
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
pymongo==4.6.3
pydantic==2.7.0
python-dotenv==1.0.1
joblib==1.4.0
